refactor: Decompose MessageHandler into command classes
The monolithic MessageHandler class has been refactored to improve modularity and maintainability.

The key changes include:
- Introduced the Command Pattern to handle individual bot commands.
- Each command (e.g., /start, /sell) is now in its own dedicated class within the 'core/handlers/Commands' directory.
- Extracted state management and automatic forward handling into their own classes: StateHandler and AutomaticForwardHandler.
- Replaced the original MessageHandler with a lean router that delegates tasks to the new specialized handlers.

This new architecture makes the code easier to understand, test, and extend.
