+++ core/handlers/Commands/CommandInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TGBot\Handlers\Commands;
+
+use TGBot\App;
+
+interface CommandInterface
+{
+    public function execute(App $app, array $message, array $parts): void;
+}
